backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
pydantic==2.7.0
pydantic-settings==2.2.1
email-validator>=2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0
croniter==6.2.2
redis==5.0.4
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
